fix: 多桌面源 + 4 主题 + Clippy 修复 + 5 agent code review 加固
## 一、原修复主线 (Multi-source Desktop / Themes / Clippy)

1. 多桌面源收纳 — 识别 Windows 三类合法桌面源 (User / Public / OneDrive)
   修复截图报告的 `Security: cannot add item outside the Desktop directory:
   C:\Users\Public\Desktop\foobar2000 plus.lnk`,信任边界不扩大。
2. 4 个 aha 主题 — brutalism / terminal / cyberpunk / editorial,
   配 is_light / font_family / border_width / effect 四个可选主题字段
   + theme-effects.css 特效层,修复 isLight 硬编码 (用 sRGB 亮度推断)。
3. Clippy 修复 — HWND_BOTTOM 保留 sentinel (#[allow] + 注释,拒绝 clippy
   危险机械建议) / is_null() / 分隔符规范化 / 删除未使用 import /
   去除冗余 as u32 / .gitignore 补 src-tauri/binaries/。

## 二、code review 加固 (5 sonnet agent 并行 → 10 issues 全修)

- CRITICAL: file_ops.rs 路径前缀缺分隔符边界 (Desktop 会绕过到
  DesktopBackup\evil.exe) → is_inside() 严格 == 或 <root>\... 匹配
- HIGH: desktop_watcher.rs 多源 watch() 全失败时静默 → attached 计数,
  为 0 返回 BentoDeskError::ConfigError
- HIGH: zones.ts setState("error", message) 把原始 JSON 写入致用户关闭
  结构化 toast 后显示 {"code":...} → structured 路径写 null
- HIGH: desktop_sources.rs SAFETY 注释错误声称 PWSTR Drop 自动释放
  (Exa 验证 windows-rs 0.58 PWSTR 是 Copy 不能有 Drop) → 重写注释
- MEDIUM: 新增 BentoDeskError::OutsideDesktop { path, allowed_sources }
  变体 + to_ipc_string() 集中 JSON envelope 包装,删除 dead message 字段
- MEDIUM: classify_source fallback 误标 → tracing::warn! 暴露异常
- MEDIUM: zones.ts as 类型断言 → in 操作符 + isStringArray 类型守卫
- MINOR: hwnd_bottom 改 const fn (编译期常量)
- MINOR: file_ops.rs 重构 validate_allowed_path_with_app_data 注入接口
  + 3 个新 test (AppData allow / AppData & Desktop 前缀碰撞 regression)
- MINOR: hidden_items.rs 加 2 个 test (尾斜杠 + 正反斜杠等价 regression)
- MINOR: App.tsx toast 内联硬编码颜色 → toast.css 类 + variables.css token
- MINOR: editorial 主题 effect: chromatic (消除 dead 类型值,匹配杂志美学)

## 三、验证结果 (全绿)

- cargo test --lib: 163/163 passed (原 155 + 8 个 code review 新 test)
- cargo clippy --all-targets: exit 0,我修改的代码零新警告
- npx tsc --noEmit: exit 0
- npx vitest run: 75/75 passed (3 suites)
- GitNexus impact 分析: setup_file_watcher_inner HIGH /
  validate_allowed_path_inner MEDIUM / is_desktop_path_with_custom LOW
- GitNexus detect_changes: 25 files / 241 symbols 在预期范围

## 四、新增文件

- src-tauri/src/desktop_sources.rs (275 行,公共 API + 7 unit test)
- src/styles/theme-effects.css (40 行,scanlines/neon/chromatic)
- src/styles/toast.css (60 行,toast 样式 token 化)

+++ src-tauri/src/drag_drop/data_object.rs
@@ -96,7 +96,7 @@ impl IDataObject_Impl for BentoDataObject_Impl {
         unsafe {
             let fmt = &*pformatetcin;
             if fmt.cfFormat == CF_HDROP.0
-                && fmt.dwAspect == DVASPECT_CONTENT.0 as u32
+                && fmt.dwAspect == DVASPECT_CONTENT.0
                 && fmt.tymed & TYMED_HGLOBAL.0 as u32 != 0
             {
                 let hglobal = self.build_hdrop()?;
